OpenCLNetwork: use the Transforms operations.
