+++ lc0/src/neural/transforms.cc
@@ -28,6 +28,34 @@
 namespace lczero 
 {
   
+  std::vector<float> Transforms::zeropad_U(const std::vector<float>& U,
+                                      const int outputs, const int channels,
+                                      const int outputs_pad,
+                                      const int channels_pad) {
+    // Fill with zeroes
+    auto Upad = std::vector<float>(WINOGRAD_TILE * outputs_pad * channels_pad);
+    
+    for(auto o = 0; o < outputs; o++) {
+      for(auto c = 0; c < channels; c++) {
+        for(auto xi = 0; xi < WINOGRAD_ALPHA; xi++){
+          for(auto nu = 0; nu < WINOGRAD_ALPHA; nu++) {
+            Upad[xi * (WINOGRAD_ALPHA * outputs_pad * channels_pad)
+                 + nu * (outputs_pad * channels_pad)
+                 + c * outputs_pad +
+                 o] =
+            U[xi * (WINOGRAD_ALPHA * outputs * channels)
+              + nu * (outputs * channels)
+              + c * outputs
+              + o];
+          }
+        }
+      }
+    }
+    
+    return Upad;
+  }
+
+  
   std::vector<float> Transforms::winograd_transform_f(const std::vector<float>& f,
                                                       const int outputs,
                                                       const int channels) {
